pyproject.toml
[project]
name = "trashbat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "matplotlib>=3.10.8",
    "pillow>=12.2.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.44.0",
]
